removed extra agent page and cell
