--- web/src/components/Agent2Cell/Agent2Cell.mock.ts
@@ -1,6 +1,0 @@
-// Define your own mock data here:
-export const standard = (/* vars, { ctx, req } */) => ({
-  agent2: {
-    id: 42,
-  },
-})
