--- web/src/components/Agent2Cell/Agent2Cell.stories.tsx
@@ -1,34 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import { Loading, Empty, Failure, Success } from './Agent2Cell'
-import { standard } from './Agent2Cell.mock'
-
-const meta: Meta = {
-  title: 'Cells/Agent2Cell',
-}
-
-export default meta
-
-export const loading: StoryObj<typeof Loading> = {
-  render: () => {
-    return Loading ? <Loading /> : <></>
-  },
-}
-
-export const empty: StoryObj<typeof Empty> = {
-  render: () => {
-    return Empty ? <Empty /> : <></>
-  },
-}
-
-export const failure: StoryObj<typeof Failure> = {
-  render: (args) => {
-    return Failure ? <Failure error={new Error('Oh no')} {...args} /> : <></>
-  },
-}
-
-export const success: StoryObj<typeof Success> = {
-  render: (args) => {
-    return Success ? <Success {...standard()} {...args} /> : <></>
-  },
-}
