--- web/src/pages/Agent2Page/Agent2Page.stories.tsx
@@ -1,13 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import Agent2Page from './Agent2Page'
-
-const meta: Meta<typeof Agent2Page> = {
-  component: Agent2Page,
-}
-
-export default meta
-
-type Story = StoryObj<typeof Agent2Page>
-
-export const Primary: Story = {}
