+++ web/src/pages/AgentPage/AgentPage.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import AgentPage from './AgentPage'
+import Agent2age from './AgentPage'
 
 const meta: Meta<typeof AgentPage> = {
   component: AgentPage,
@@ -8,6 +8,6 @@ const meta: Meta<typeof AgentPage> = {
 
 export default meta
 
-type Story = StoryObj<typeof AgentPage>
+type Story = StoryObj<typeof Agent2Page>
 
 export const Primary: Story = {}
