+++ web/src/pages/AgentPage/AgentPage.test.tsx
@@ -5,7 +5,7 @@ import AgentPage from './AgentPage'
 //   Improve this test with help from the Redwood Testing Doc:
 //   https://redwoodjs.com/docs/testing#testing-pages-layouts
 
-describe('AgentPage', () => {
+describe('Agent2Page', () => {
   it('renders successfully', () => {
     expect(() => {
       render(<AgentPage />)
